vercel_deploy_fix-002: other fixes to ui + small refactor
