+++ app/admin/components/data-table-access-control-user-roles.tsx
@@ -26,7 +26,7 @@ import {
 } from "@/components/ui/table"
 
 import { DataTablePagination } from "../components/data-table-pagination"
-import { DataTableToolbar } from "../components/data-table-toolbar"
+import { DataTableToolbar } from "../components/data-table-toolbar-access-control-user-roles"
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
