+++ app/admin/knowledgebase-details/knowledgebase-table.tsx
@@ -4,8 +4,8 @@ import { Metadata } from "next"
 import Image from "next/image"
 import { z } from "zod"
 
-import { columns } from "@/app/admin/components/columns"
-import { DataTable } from "@/app/admin/components/data-table"
+import { columns } from "@/app/admin/components/columns-knowledgebase-details"
+import { DataTable } from "@/app/admin/components/data-table-knowledgebase-details"
 import { UserNav } from "@/app/admin/components/user-nav"
 import { taskSchema } from "@/app/admin/data/schema"
 import moment from 'moment'
@@ -15,12 +15,12 @@ import useSWR from "swr";
 function KnowledgeBaseTable() {
   const fetcher = (url) => fetch(url, {cache: "no-store"}).then((res) => res.json());
   const { data, error, isLoading, mutate } = useSWR(
-    '/api/admin/get-vector-log-data',
+    '/api/admin/knowledgebase/get-vector-log-data',
     fetcher,
-    { refreshInterval: 1000 }
+    { refreshInterval: 0 }
   );
   if (isLoading) return <p>Loading...</p>
-  if (!data) return <p>No vector log data</p>
+  if (!data) return <p>No data!</p>
