+++ app/admin/knowledgebase.tsx
@@ -59,7 +59,7 @@ export function KnowledgeBase(vector_data_log: JSON) {
     setReady(false);
 
     let payload = { "text": inputText }
-    const response = await fetch('/api/admin/embedding-text-insert', {
+    const response = await fetch('/api/admin/knowledgebase/embedding-text-insert', {
       method: 'POST',
       headers: {},
       body: JSON.stringify(payload)
@@ -90,7 +90,7 @@ export function KnowledgeBase(vector_data_log: JSON) {
     formData.append('file_name', FileState.selectedFile.name);
     formData.append('file', FileState.selectedFile);
 
-    const response = await fetch('/api/admin/markdown-file-upload', {
+    const response = await fetch('/api/admin/knowledgebase/markdown-file-upload', {
       method: 'POST',
       body: formData
     }
