+++ app/api/admin/knowledgebase/get-vector-log-data/route.ts
